fix(api): remaining money columns to BigInteger + debt date schema
Audit revealed 4 monetary columns missed in the previous BigInteger
sweep. Fixed and migration applied to staging and production:

- finance_snapshots.net_savings
- finance_snapshots.total_debt
- finance_snapshots.total_savings
- debt_payments.interest

Also fixes a pre-existing bug exposed during E2E validation: the
DebtCreate / DebtUpdate / DebtResponse Pydantic schemas declared
start_date and expected_end_date as 'str' but the SQLAlchemy model has
Date columns. asyncpg requires actual datetime.date objects and was
raising 'str object has no attribute toordinal'. Switched to date type
so Pydantic auto-validates ISO strings into date objects.

E2E validated locally: created a debt with original_amount=500 million
Kz (50 billion centavos) successfully — the exact payload that was
crashing in production. Verified stored value is correct in the DB.

+++ apps/api/app/models/debt.py
@@ -81,7 +81,7 @@ class DebtPayment(BaseModel):
     )
     amount: Mapped[int] = mapped_column(BigInteger)  # centavos
     principal: Mapped[int | None] = mapped_column(BigInteger)
-    interest: Mapped[int | None] = mapped_column(Integer)
+    interest: Mapped[int | None] = mapped_column(BigInteger)
     payment_date: Mapped[date] = mapped_column(Date)
     transaction_id: Mapped[uuid.UUID | None] = mapped_column(
         PG_UUID(as_uuid=True), ForeignKey("transactions.id")
